+++ packages/editor/js/extensions/libs/background/styles.js
@@ -43,6 +43,7 @@ export const BackgroundStyles = ({
 		state,
 		clientId,
 		blockName,
+		currentBlock,
 		attributes: currentBlockAttributes,
 	};
 	const sharedParams = {