+++ packages/editor/js/extensions/libs/typography/styles.js
@@ -61,6 +61,7 @@ export function TypographyStyles({
 		state,
 		clientId,
 		blockName,
+		currentBlock,
 		attributes: currentBlockAttributes,
 	};