+++ packages/editor/php/StyleEngine.php
@@ -229,9 +229,11 @@ final class StyleEngine {
 				array_filter(
 					array_map(
                         function( array $stateSettings, string $state): array {
+							$this->pseudo_state = $state;
+
 							return array_map(
-                                function ( $breakpointSettings, string $breakpoint) use ( $state): string {
-                                    return $this->prepareBreakpointStyles($breakpoint, $breakpointSettings['attributes'], $state);
+                                function ( $breakpointSettings, string $breakpoint): string {
+                                    return $this->prepareBreakpointStyles($breakpoint, $breakpointSettings['attributes']);
                                 },
                                 $stateSettings['breakpoints'],
                                 array_keys($stateSettings['breakpoints'])
@@ -259,11 +261,10 @@ final class StyleEngine {
 	 *
 	 * @param string $breakpoint The breakpoint type.
 	 * @param array  $settings The current breakpoint settings.
-	 * @param string $state The block state. Default is 'normal'.
 	 *
 	 * @return string The generated css rule for current breakpoint.
 	 */
-	protected function prepareBreakpointStyles( string $breakpoint, array $settings, string $state = 'normal' ): string {
+	protected function prepareBreakpointStyles( string $breakpoint, array $settings ): string {
 
 		// Get css media queries.
 		$mediaQueries = blockera_get_css_media_queries($this->breakpoints['list']);
@@ -278,7 +279,7 @@ final class StyleEngine {
 		$this->breakpoint = $breakpoint;
 
 		// We should just prepare normal state styles because not exists any other states.
-		$state_css_rules = $this->prepareStateStyles($state, $settings);
+		$state_css_rules = $this->prepareStateStyles($settings);
 
 		// Exclude empty css rules.
 		if ( empty( $state_css_rules ) ) {
@@ -372,14 +373,11 @@ final class StyleEngine {
 	/**
 	 * Preparing css of current state settings.
 	 *
-	 * @param string $state The state name (as pseudo class in css).
-	 * @param array  $settings the breakpoint current state settings.
+	 * @param array $settings the breakpoint current state settings.
 	 *
 	 * @return array The state css rules.
 	 */
-	protected function prepareStateStyles( string $state, array $settings ): array {
-
-		$this->pseudo_state = $state;
+	protected function prepareStateStyles( array $settings ): array {
 
 		$block_css = array_map(
             function ( $settings, string $id): array {